requirements.txt
google-generativeai
python-dotenv
mcp